feat(download): add opa_no_oci flag to build without containerd
